fix: update_product - sync service_products + media processing/deletion + media URL normalization
Backend (products_management.rs):
- BUG 3: Sync product_data to service_products table on update
- BUG 4: Process new base64 media uploads during product edit
- BUG 5: Delete old media entries removed by user during edit
- Clean base64 data from JSONB before saving
- Handle GENERATED ALWAYS columns (product_name, product_price)

Mobile (MesProduitsScreen.tsx):
- Fix: Convert storage paths (uploads/...) to full URLs via buildMediaUrl
- Applied to both handleEditProduct and handleDuplicateProduct
- Ensures existing media display correctly in edit/duplicate forms

+++ backend/src/routes/products_management.rs
@@ -1064,7 +1064,7 @@ pub async fn update_product(
             match processor
                 .process_media_batch(
                     service_id_i32,
-                    Some(payload.product_index as u32),
+                    Some(payload.product_index as usize),
                     media_items,
                 )
                 .await
